contests

+++ app/Http/Controllers/ContestController.php
@@ -8,10 +8,12 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Requests;
 use App\ProgrammingLanguage;
+use App\User;
 
 class ContestController extends Controller
 {
-    public function index(Request $request) {
+    public function index(Request $request)
+    {
         $orderBySession = \Session::get('orderBy', 'updated_at');
         $orderBy = $request->input('order', $orderBySession);
 
@@ -20,7 +22,6 @@ class ContestController extends Controller
 
         $page = $request->input('page');
         $query = $request->input('query', '');
-
 
 
         if (!in_array($orderBy, Contest::sortable())) {
@@ -34,10 +35,17 @@ class ContestController extends Controller
         \Session::put('orderBy', $orderBy);
         \Session::put('orderDir', $orderDir);
 
-
-
-        $contests = Contest::orderBy($orderBy, $orderDir)
-            ->paginate(10);
+        $contests = Contest::orderBy($orderBy, $orderDir);
+        if (Auth::check() && Auth::user()->hasRole(User::ROLE_TEACHER)) {
+            $contests = $contests->where('user_id', Auth::user()->id);
+        } elseif (Auth::check() && Auth::user()->hasRole(User::ROLE_USER)) {
+            $contests = $contests->whereHas('users', function($query){
+                $query->where('user_id', Auth::user()->id);
+            })->orWhere('labs', true);
+        } elseif (!Auth::check() || Auth::user()->hasRole(User::ROLE_LOW_USER)) {
+            $contests = $contests->where('labs', true);
+        }
+        $contests = $contests->paginate(10);
 
         return view('contests.list')->with([
             'contests' => $contests,
@@ -47,26 +55,30 @@ class ContestController extends Controller
             'query' => $query
         ]);
     }
+
     public function showForm(Request $request, $id = null)
     {
         $contest = ($id ? Contest::findOrFail($id) : new Contest());
-        if ($id) {
-            $title = 'Edit Contest';
-            $participants = $contest->users()->user()->get();
-        } else {
-            $title = 'Create Contest';
-            $participants = null;
-        }
+        if ($id && $contest->currentUserAllowedEdit() || !$id) {
+            if ($id) {
+                $title = 'Edit Contest';
+                $participants = $contest->users()->user()->get();
+            } else {
+                $title = 'Create Contest';
+                $participants = collect();
+            }
 
-        $students = Auth::user()->students()->get()->diff($participants);
-        
-        return view('contests.form')->with([
-            'contest' => $contest,
-            'title' => $title,
-            'students' => $students,
-            'participants' => $participants,
-            'programming_languages' => ProgrammingLanguage::orderBy('name', 'desc')->get()
-        ]);
+            $students = Auth::user()->students()->get()->diff($participants);
+
+            return view('contests.form')->with([
+                'contest' => $contest,
+                'title' => $title,
+                'students' => $students,
+                'participants' => $participants,
+                'programming_languages' => ProgrammingLanguage::orderBy('name', 'desc')->get()
+            ]);
+        }
+        return redirect()->route('contests::list');
     }
 
     /**
@@ -80,39 +92,58 @@ class ContestController extends Controller
     public function edit(Request $request, $id = null)
     {
         $contest = (!$id ?: Contest::findOrFail($id));
-        $fillData = [
-            'name' => $request->get('name'),
-            'description' => $request->get('description'),
-            'user_id' => Auth::user()->id,
-            'start_date' => $request->get('start_date'),
-            'end_date' => $request->get('end_date'),
-            'is_active' => $request->get('is_active'),
-            'is_standings_active' => $request->get('is_standings_active'),
-        ];
+        if ($id && $contest->currentUserAllowedEdit() || !$id) {
+            $fillData = [
+                'name' => $request->get('name'),
+                'description' => $request->get('description'),
+                'start_date' => $request->get('start_date'),
+                'end_date' => $request->get('end_date'),
+                'user_id' => Auth::user()->id,
+                'is_active' => $request->get('is_active'),
+                'is_standings_active' => $request->get('is_standings_active'),
+                'show_max' => $request->get('show_max'),
+                'labs' => $request->get('labs'),
+            ];
 
-        $this->validate($request, Contest::getValidationRules());
+            $this->validate($request, Contest::getValidationRules());
 
-        if ($id) {
-            $contest->fill($fillData);
-        } else {
-            $contest = Contest::create($fillData);
+
+
+            if ($id) {
+                $contest->fill($fillData);
+            } else {
+                $contest = Contest::create($fillData);
+            }
+
+            $contest->programming_languages()->detach();
+            $contest->programming_languages()->attach($request->get('programming_languages'));
+
+
+            foreach (Auth::user()->students as $student) {
+                $contest->users()->detach($student->id);
+            }
+
+            $contest->users()->attach($request->get('participants'));
+
+            $contest->save();
+
+            \Session::flash('alert-success', 'The contest was successfully saved');
         }
-
-        $contest->programming_languages()->detach();
-        $contest->programming_languages()->attach($request->get('programming_languages'));
-
-
-
-        foreach(Auth::user()->students as $student) {
-            $contest->users()->detach($student->id);
-        }
-
-        $contest->users()->attach($request->get('participants'));
-        
-        $contest->save();
-
-        \Session::flash('alert-success', 'The contest was successfully saved');
-        return redirect()->route('teacherOnly::contests::list');
+        return redirect()->route('contests::list');
     }
 
+    public function hide(Request $request, $id)
+    {
+
+        $contest = Contest::findOrFail($id);
+        if ($contest->currentUserAllowedEdit()) {
+            $contest->hide();
+            $contest->save();
+        }
+        return redirect()->route('contests::list');
+    }
+
+    public function single(Request $request, $id) {
+        return View('contests.single')->with(['contest' => Contest::findOrFail($id)]);
+    }
 }
